Fix warning codes in Easysetup
Reviewed-on: https://gerrit.iotivity.org/gerrit/12309
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
(cherry picked from commit 6b311610a9aec78c4da0f5520ed928244719e3ff)
Reviewed-on: https://gerrit.iotivity.org/gerrit/12533

+++ service/easy-setup/mediator/richsdk/src/EnrolleeSecurity.cpp
@@ -159,7 +159,7 @@ namespace OIC
             else if (pOwnedDevList.size())
             {
                 OIC_LOG_V(DEBUG, ENROLEE_SECURITY_TAG, "Found owned devices. Count =%d",
-                        pOwnedDevList.size());
+                        (int)pOwnedDevList.size());
                 std::shared_ptr< OC::OCSecureResource > ownedDevice =
                     findEnrolleeSecurityResource(pOwnedDevList);
 
@@ -182,7 +182,7 @@ namespace OIC
             else if (pUnownedDevList.size())
             {
                 OIC_LOG_V(DEBUG, ENROLEE_SECURITY_TAG, "Found Unowned devices. Count =%d",
-                        pUnownedDevList.size());
+                        (int)pUnownedDevList.size());
 
                 m_unownedDevice = findEnrolleeSecurityResource(pUnownedDevList);
                 if (m_unownedDevice)
@@ -330,7 +330,7 @@ namespace OIC
             else if (pOwnedDevList.size())
             {
                 OIC_LOG_V(DEBUG, ENROLEE_SECURITY_TAG, "Found owned devices. Count =%d",
-                        pOwnedDevList.size());
+                        (int)pOwnedDevList.size());
                 ownedDevice = findEnrolleeSecurityResource(pOwnedDevList);
 
                 if (!ownedDevice)
